+++ src/main/java/pl/asie/computronics/integration/mfr/DriverDeepStorageUnit.java
@@ -3,54 +3,15 @@ package pl.asie.computronics.integration.mfr;
 import dan200.computercraft.api.lua.ILuaContext;
 import dan200.computercraft.api.lua.LuaException;
 import dan200.computercraft.api.peripheral.IComputerAccess;
-import li.cil.oc.api.machine.Arguments;
-import li.cil.oc.api.machine.Callback;
-import li.cil.oc.api.machine.Context;
-import li.cil.oc.api.network.ManagedEnvironment;
-import li.cil.oc.api.prefab.DriverTileEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import pl.asie.computronics.api.multiperipheral.IMultiPeripheral;
 import pl.asie.computronics.integration.CCMultiPeripheral;
-import pl.asie.computronics.integration.ManagedEnvironmentOCTile;
 import pl.asie.computronics.reference.Names;
 import powercrystals.minefactoryreloaded.api.IDeepStorageUnit;
 
 public class DriverDeepStorageUnit {
-	public static class OCDriver extends DriverTileEntity {
-		public class InternalManagedEnvironment extends ManagedEnvironmentOCTile<IDeepStorageUnit> {
-			public InternalManagedEnvironment(IDeepStorageUnit tile, String name) {
-				super(tile, name);
-			}
-
-			@Callback()
-			public Object[] getStoredItem(Context c, Arguments a) {
-				return new Object[] { tile.getStoredItemType() };
-			}
-
-			@Callback(direct = true)
-			public Object[] isLocked(Context c, Arguments a) {
-				ItemStack is = tile.getStoredItemType();
-				return new Object[] { (is != null && is.stackSize == 0) };
-			}
-
-			@Callback(direct = true)
-			public Object[] getMaxItemCount(Context c, Arguments a) {
-				return new Object[] { tile.getMaxStoredCount() };
-			}
-		}
-
-		@Override
-		public Class<?> getTileEntityClass() {
-			return IDeepStorageUnit.class;
-		}
-
-		@Override
-		public ManagedEnvironment createEnvironment(World world, int x, int y, int z) {
-			return new InternalManagedEnvironment((IDeepStorageUnit) world.getTileEntity(x, y, z), Names.MFR_DSU);
-		}
-	}
 
 	public static class CCDriver extends CCMultiPeripheral<IDeepStorageUnit> {
 
@@ -89,7 +50,7 @@ public class DriverDeepStorageUnit {
 					return new Object[] { null };
 				}
 			}
-			switch(method){
+			switch(method) {
 				case 0:
 					return new Object[] { is.getUnlocalizedName() };
 				case 1: