+++ src/main/java/pl/asie/computronics/integration/waila/IntegrationWaila.java
@@ -8,6 +8,8 @@ import pl.asie.computronics.reference.Mods;
 
 public class IntegrationWaila {
 	public static void register(IWailaRegistrar reg) {
+		WailaProviders.initialize();
+
 		WailaComputronics provider = new WailaComputronics();
 		reg.registerBodyProvider(provider, BlockPeripheral.class);
 		reg.registerNBTProvider(provider, BlockPeripheral.class);