+++ src/main/java/pl/asie/computronics/integration/waila/providers/WailaPeripheral.java
@@ -2,9 +2,11 @@ package pl.asie.computronics.integration.waila.providers;
 
 import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.Optional;
+import li.cil.oc.api.network.Environment;
 import li.cil.oc.api.network.Node;
 import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
+import nedocomputers.api.INedoPeripheral;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
@@ -12,8 +14,8 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import pl.asie.computronics.integration.waila.ConfigValues;
 import pl.asie.computronics.reference.Mods;
-import pl.asie.computronics.tile.TileEntityPeripheralBase;
 import pl.asie.computronics.util.StringUtil;
+import pl.asie.computronics.util.tile.IComputronicsPeripheral;
 
 import java.util.List;
 
@@ -56,20 +58,23 @@ public class WailaPeripheral extends ComputronicsWailaProvider {
 
 	@Override
 	public NBTTagCompound getNBTData(EntityPlayerMP player, TileEntity te, NBTTagCompound tag, World world, int x, int y, int z) {
-		if(te != null && te instanceof TileEntityPeripheralBase) {
-			TileEntityPeripheralBase tile = (TileEntityPeripheralBase) te;
+		if(te != null && te instanceof IComputronicsPeripheral) {
 			if(Loader.isModLoaded(Mods.OpenComputers)) {
-				tag = getNBTData_OC(tile, tag);
+				tag = getNBTData_OC(te, tag);
 			}
 			if(Loader.isModLoaded(Mods.NedoComputers)) {
-				tag = getNBTData_NC(tile, tag);
+				tag = getNBTData_NC(te, tag);
 			}
 		}
 		return tag;
 	}
 
 	@Optional.Method(modid = Mods.OpenComputers)
-	public NBTTagCompound getNBTData_OC(TileEntityPeripheralBase tile, NBTTagCompound tag) {
+	public NBTTagCompound getNBTData_OC(TileEntity te, NBTTagCompound tag) {
+		if(!(te instanceof Environment)) {
+			return tag;
+		}
+		Environment tile = ((Environment) te);
 		Node node = tile.node();
 		if(node != null && node.host() == tile) {
 			final NBTTagCompound nodeNbt = new NBTTagCompound();
@@ -80,7 +85,11 @@ public class WailaPeripheral extends ComputronicsWailaProvider {
 	}
 
 	@Optional.Method(modid = Mods.NedoComputers)
-	private NBTTagCompound getNBTData_NC(TileEntityPeripheralBase tile, NBTTagCompound tag) {
+	private NBTTagCompound getNBTData_NC(TileEntity te, NBTTagCompound tag) {
+		if(!(te instanceof INedoPeripheral)) {
+			return tag;
+		}
+		INedoPeripheral tile = ((INedoPeripheral) te);
 		if(tile.getBusId() != 0) {
 			tag.setShort("nc:bus", (short) tile.getBusId());
 		}