+++ src/main/java/pl/asie/computronics/tile/TileEntityPeripheralBase.java
@@ -5,6 +5,7 @@ import cpw.mods.fml.common.Optional;
 import dan200.computercraft.api.peripheral.IComputerAccess;
 import dan200.computercraft.api.peripheral.IPeripheral;
 import li.cil.oc.api.Network;
+import li.cil.oc.api.network.BlacklistedPeripheral;
 import li.cil.oc.api.network.Environment;
 import li.cil.oc.api.network.Message;
 import li.cil.oc.api.network.Node;
@@ -12,8 +13,9 @@ import li.cil.oc.api.network.Visibility;
 import nedocomputers.api.INedoPeripheral;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
-import pl.asie.computronics.cc.IComputronicsPeripheral;
+import pl.asie.computronics.api.multiperipheral.IMultiPeripheral;
 import pl.asie.computronics.reference.Mods;
+import pl.asie.computronics.util.tile.IComputronicsPeripheral;
 import pl.asie.lib.tile.TileMachine;
 
 import java.util.ArrayList;
@@ -26,10 +28,12 @@ import java.util.ArrayList;
 
 @Optional.InterfaceList({
 	@Optional.Interface(iface = "li.cil.oc.api.network.Environment", modid = Mods.OpenComputers),
-	@Optional.Interface(iface = "pl.asie.computronics.cc.IComputronicsPeripheral", modid = Mods.ComputerCraft),
+	@Optional.Interface(iface = "li.cil.oc.api.network.BlacklistedPeripheral", modid = Mods.OpenComputers),
+	@Optional.Interface(iface = "pl.asie.computronics.api.multiperipheral.IMultiPeripheral", modid = Mods.ComputerCraft),
 	@Optional.Interface(iface = "nedocomputers.api.INedoPeripheral", modid = Mods.NedoComputers)
 })
-public abstract class TileEntityPeripheralBase extends TileMachine implements Environment, IComputronicsPeripheral, INedoPeripheral {
+public abstract class TileEntityPeripheralBase extends TileMachine implements Environment,
+	IMultiPeripheral, IComputronicsPeripheral, INedoPeripheral, BlacklistedPeripheral {
 	protected String peripheralName;
 
 	public TileEntityPeripheralBase(String name) {
@@ -88,6 +92,12 @@ public abstract class TileEntityPeripheralBase extends TileMachine implements En
 	@Override
 	@Optional.Method(modid = Mods.OpenComputers)
 	public void onMessage(final Message message) {
+	}
+
+	@Override
+	@Optional.Method(modid = Mods.OpenComputers)
+	public boolean isPeripheralBlacklisted() {
+		return true;
 	}
 
 	@Override